Databases wip
Add database support for PostgreSQL and Sqlite3, refactor stuff to modules. Wip.
